Moving classes from ModeOfInheritanceFilter to the package level.
