+++ jannovar-filter/src/main/java/de/charite/compbio/jannovar/filter/ModeOfInheritanceFilter.java
@@ -10,11 +10,9 @@ import java.util.HashMap;
 import java.util.Map;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 
 import de.charite.compbio.jannovar.impl.intervals.Interval;
 import de.charite.compbio.jannovar.impl.intervals.IntervalArray;
-import de.charite.compbio.jannovar.impl.intervals.IntervalEndExtractor;
 import de.charite.compbio.jannovar.io.Chromosome;
 import de.charite.compbio.jannovar.io.JannovarData;
 import de.charite.compbio.jannovar.io.ReferenceDictionary;
@@ -336,217 +334,6 @@ public class ModeOfInheritanceFilter implements VariantContextFilter {
 
 		// mark gene as done
 		activeGenes.remove(gene);
-	}
-
-	/**
-	 * Store {@link FlaggedVariant} and a counter.
-	 */
-	private static class FlaggedVariantCounter {
-		public final FlaggedVariant var;
-		public final GenomeChange change;
-		public int count;
-
-		FlaggedVariantCounter(FlaggedVariant var, GenomeChange change, int count) {
-			this.var = var;
-			this.change = change;
-			this.count = count;
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result + ((var == null) ? 0 : var.hashCode());
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj)
-				return true;
-			if (obj == null)
-				return false;
-			if (getClass() != obj.getClass())
-				return false;
-			FlaggedVariantCounter other = (FlaggedVariantCounter) obj;
-			if (var == null) {
-				if (other.var != null)
-					return false;
-			} else if (!var.equals(other.var))
-				return false;
-			return true;
-		}
-
-	}
-
-	/**
-	 * Simple representation of a gene.
-	 *
-	 * Genes are identified by their name, {@link #equals} and {@link #hashCode} only consider the field {@link #name}!
-	 */
-	private static class Gene {
-		public final String name;
-		public final ImmutableList<TranscriptModel> transcripts;
-		public final ReferenceDictionary refDict;
-		public final GenomeInterval region;
-
-		public Gene(ReferenceDictionary refDict, String name, ImmutableList<TranscriptModel> transcripts) {
-			this.refDict = refDict;
-			this.name = name;
-			this.transcripts = transcripts;
-			this.region = buildGeneRegion();
-		}
-
-		/**
-		 * @return {@link GenomeInterval} of this gene (smallest begin and largest end position of all transcripts).
-		 */
-		private GenomeInterval buildGeneRegion() {
-			if (transcripts.isEmpty())
-				return null;
-
-			GenomeInterval region = transcripts.get(0).txRegion.withStrand('+');
-			for (TranscriptModel tm : transcripts)
-				region = mergeRegions(region, tm.txRegion);
-			return region;
-		}
-
-		/**
-		 * @return {@link GenomeInterval} from the smaller begin to the larger end position of <code>lhs</code> and
-		 *         <code>rhs</code>.
-		 */
-		private GenomeInterval mergeRegions(GenomeInterval lhs, GenomeInterval rhs) {
-			lhs = lhs.withStrand('+');
-			rhs = rhs.withStrand('+');
-			return new GenomeInterval(lhs.getGenomeBeginPos().refDict, '+', lhs.getGenomeBeginPos().chr, Math.min(
-					lhs.beginPos, rhs.beginPos), Math.max(lhs.endPos, rhs.endPos));
-		}
-
-		@Override
-		public String toString() {
-			return this.name + ":" + this.region.toString();
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result + ((name == null) ? 0 : name.hashCode());
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj)
-				return true;
-			if (obj == null)
-				return false;
-			if (getClass() != obj.getClass())
-				return false;
-			Gene other = (Gene) obj;
-			if (name == null) {
-				if (other.name != null)
-					return false;
-			} else if (!name.equals(other.name)) {
-				return false;
-			}
-			return true;
-		}
-	}
-
-	/**
-	 * Builder for {@link Gene}.
-	 */
-	private static class GeneBuilder {
-		private final ReferenceDictionary refDict;
-		private String name = null;
-		private final ImmutableList.Builder<TranscriptModel> builder = new ImmutableList.Builder<TranscriptModel>();
-		private final ArrayList<TranscriptModel> tmpModels = new ArrayList<TranscriptModel>();
-
-		public GeneBuilder(ReferenceDictionary refDict, String name) {
-			this.refDict = refDict;
-			this.name = name;
-		}
-
-		public void addTranscriptModel(TranscriptModel tm) {
-			if (tmpModels.isEmpty()) { // always add first
-				System.err.println("Adding first transcript " + tm.accession + " " + tm.txRegion + " to gene " + name);
-				builder.add(tm);
-				tmpModels.add(tm);
-				return;
-			}
-
-			// Transcript must be within 10kbp of a previously seen one. Otherwise, we get too large genes from RNA
-			// transcript matches.
-			final int MORE_PADDING = 10000;
-			final GenomeInterval tmRegion = tm.txRegion.withMorePadding(MORE_PADDING);
-			for (TranscriptModel model : tmpModels)
-				if (model.txRegion.overlapsWith(tmRegion)) {
-					System.err.println("Adding transcript " + tm.accession + " " + tm.txRegion + " to gene " + name);
-					builder.add(tm);
-					tmpModels.add(tm);
-					return;
-				}
-			System.err.println("Transcript " + tm.accession + " " + tm.txRegion
-					+ " does not fit to previous transcripts of " + name);
-		}
-
-		public Gene build() {
-			final Gene gene = new Gene(refDict, name, builder.build());
-			System.err.println("CREATING GENE\t" + gene.toString() + "\tLENGTH\t" + gene.region.length());
-			return gene;
-		}
-
-		public String getName() {
-			return name;
-		}
-
-		public void setName(String name) {
-			this.name = name;
-		}
-	}
-
-	/**
-	 * Extraction of interval ends of {@link Gene} objects.
-	 */
-	private static class GeneIntervalEndExtractor implements IntervalEndExtractor<Gene> {
-		@Override
-		public int getBegin(Gene gene) {
-			return gene.region.beginPos;
-		}
-
-		@Override
-		public int getEnd(Gene gene) {
-			return gene.region.endPos;
-		}
-	}
-
-	/**
-	 * List of genes, accessible through an interval tree.
-	 */
-	private static class GeneList {
-		/** overall gene list */
-		public final ImmutableList<Gene> genes;
-		/** map from numeric chromosome id to interval tree of genes */
-		public final ImmutableMap<Integer, IntervalArray<Gene>> gIntervalTree;
-
-		public GeneList(ImmutableList<Gene> genes) {
-			this.genes = genes;
-			this.gIntervalTree = buildIntervalTree();
-		}
-
-		private ImmutableMap<Integer, IntervalArray<Gene>> buildIntervalTree() {
-			HashMap<Integer, ArrayList<Gene>> chrToGene = new HashMap<Integer, ArrayList<Gene>>();
-			for (Gene gene : genes) {
-				if (!chrToGene.containsKey(gene.region.chr))
-					chrToGene.put(gene.region.chr, new ArrayList<Gene>());
-				chrToGene.get(gene.region.chr).add(gene);
-			}
-
-			ImmutableMap.Builder<Integer, IntervalArray<Gene>> builder = new ImmutableMap.Builder<Integer, IntervalArray<Gene>>();
-			for (Map.Entry<Integer, ArrayList<Gene>> entry : chrToGene.entrySet())
-				builder.put(entry.getKey(), new IntervalArray<Gene>(entry.getValue(), new GeneIntervalEndExtractor()));
-			return builder.build();
-		}
 	}
 
 }
